crawler/requirements.txt
fast-flights>=2.2
supabase>=2.28
python-dotenv>=1.1
